ARM: omap5: Enable PCIe for Elesar Titanium
For the Elesar Titanium board set the PCIe mux to be 2x single lane mode,
set the physical coding sublayer delay to nominal and turn on the low
jitter clock buffer to output the 100MHz reference clock.

This setup is necessary for PCIe to work in Linux when it loads.